requirements.txt
django
djangorestframework
django-filter
djoser
Pillow
Faker